Switch to Zuul's regular expression for `Depends-On` detection
Since `Depends-On` is ultimately for Zuul (the service), there is no
point in being stricter or less strict that Zuul is. Otherwise, the UI
would not reflect what Zuul can pick up or fails to pick up. This would
be confusing for users.

For reference, Zuul's regular expression for `Depends-On` detection is
here:

  https://opendev.org/zuul/zuul/src/commit/041e251b7c3d76f7cc89e491f2ed246b87b1cd9c/zuul/driver/gerrit/gerritconnection.py#L477

Change-Id: Idf421a6e1a29e21883540d49cf5ccc8190ffec17