Expose ChangeInfos instead of Change-Ids
As not all dependencies need to be available on the server, we need to
split dependencies into available and missing ones. For the available
ones, we offer ChangeInfos, while for the missing ones we can still
only offer Change-Ids.

With this change, the UI shows project, branch and subject of the
changes and thereby better blends in with the other list of changes.

Change-Id: I7cfdba5c2942341b6bbb3292f91b35abf5a1603f